+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Mux.hs
@@ -165,7 +165,7 @@ demo chain0 updates delay = do
                                        }
 
     clientAsync <- async $ do
-      clientMux <- Mx.new (toMiniProtocolInfos consumerApp)
+      clientMux <- Mx.new (toMiniProtocolInfos (\_ _ -> Nothing) consumerApp)
       let initCtx = MinimalInitiatorContext (ConnectionId "consumer" "producer")
       resOps <- sequence
         [ Mx.runMiniProtocol
@@ -190,7 +190,7 @@ demo chain0 updates delay = do
         wait aid
 
     serverAsync <- async $ do
-      serverMux <- Mx.new (toMiniProtocolInfos producerApp)
+      serverMux <- Mx.new (toMiniProtocolInfos (\_ _ -> Nothing) producerApp)
       let respCtx = ResponderContext (ConnectionId "producer" "consumer")
       resOps <- sequence
         [ Mx.runMiniProtocol